Fix uint when uchar is sufficient

+++ code/render/coregraphics/accelerationstructure.h
@@ -65,7 +65,7 @@ struct BlasInstanceCreateInfo
     CoreGraphics::BlasId blas;
     Math::mat4 transform;
     uint instanceIndex;         // Readable in the shader as gl_InstanceCustomIndexKHR
-    uint mask;                  // 8 bit visibility mask
+    uchar mask;                  // 8 bit visibility mask
     uint shaderOffset;          // Offset into the shader binding table
     BlasInstanceFlags flags;
 };
